move thunks back to the same file to avoid circular dependencies
