+++ src/pageEditor/panes/EditorPane.tsx
@@ -16,7 +16,10 @@
  */
 
 import React, { useEffect } from "react";
-import { actions as editorActions } from "@/pageEditor/slices/editorSlice";
+import {
+  actions,
+  actions as editorActions,
+} from "@/pageEditor/slices/editorSlice";
 import { useDispatch, useSelector } from "react-redux";
 import { useDebouncedCallback } from "use-debounce";
 import ErrorBoundary from "@/components/ErrorBoundary";
@@ -30,7 +33,6 @@ import {
   selectActiveElement,
   selectSelectionSeq,
 } from "@/pageEditor/slices/editorSelectors";
-import { checkActiveElementAvailability } from "@/pageEditor/slices/editorThunks";
 
 // CHANGE_DETECT_DELAY_MILLIS should be low enough so that sidebar gets updated in a reasonable amount of time, but
 // high enough that there isn't an entry lag in the page editor
@@ -47,7 +49,7 @@ const EditorPaneContent: React.VoidFunctionComponent<{
   const syncReduxState = useDebouncedCallback(
     (values: FormState) => {
       dispatch(editorActions.editElement(values));
-      dispatch(checkActiveElementAvailability());
+      dispatch(actions.checkActiveElementAvailability());
     },
     REDUX_SYNC_WAIT_MILLIS,
     { trailing: true, leading: false }
