+++ src/pageEditor/sidebar/DynamicEntry.tsx
@@ -47,7 +47,6 @@ import useSaveExtension from "@/pageEditor/hooks/useSaveExtension";
 import useResetExtension from "@/pageEditor/hooks/useResetExtension";
 import useRemoveExtension from "@/pageEditor/hooks/useRemoveExtension";
 import useSaveRecipe from "@/pageEditor/hooks/useSaveRecipe";
-import { cloneActiveExtension } from "@/pageEditor/slices/editorThunks";
 
 type DynamicEntryProps = {
   extension: FormState;
@@ -107,7 +106,7 @@ const DynamicEntry: React.FunctionComponent<DynamicEntryProps> = ({
   const onRemove = async () => removeExtension({ extensionId: extension.uuid });
 
   const onClone = async () => {
-    dispatch(cloneActiveExtension());
+    dispatch(actions.cloneActiveExtension());
   };
 
   return (
